# packages/backend-defaults/src/openapi.yaml
openapi: 3.1.0
info:
  title: Roadie Backend API
  version: 1.0.0
  description: >
    Comprehensive API specification for the Roadie backend,
    covering catalog workflows, executions, integrations, datastore,
    secrets, feature flags, permissions, and AI.
servers:
  - url: /api
    description: Backend API base path
security:
  - ServiceToken: []
...
paths:
...
  /catalog-workflow/executions/{id}/stream:
    parameters:
      - $ref: '#/components/parameters/ExecutionId'
    get:
      operationId: streamExecution
      tags: [catalog-workflow-executions]
      summary: Stream execution events via SSE
      responses:
        '200':
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /catalog-workflow/nodes:
    get:
      operationId: listNodeTypes
      tags: [catalog-workflow-nodes]
      summary: List available node type definitions
      parameters:
        - name: workflowType
          in: query
          schema:
            $ref: '#/components/schemas/WorkflowType'
      responses:
        '200':
          description: Node definitions
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/NodeTypeDefinition'
  /catalog-workflow/nodes/{type}:
    parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getNodeType
      tags: [catalog-workflow-nodes]
      summary: Get a single node type definition
      responses:
        '200':
          description: Node definition
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/NodeTypeDefinition'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================================================
  # CATALOG-WORKFLOW — Graph Layouts
  # ==========================================================================
...
  /catalog-workflow/graph-layouts/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getGraphLayout
      tags: [catalog-workflow-graph-layouts]
      summary: Get a graph layout by name
      responses:
        '200':
          description: Layout detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GraphLayout'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: upsertGraphLayout
      tags: [catalog-workflow-graph-layouts]
      summary: Create or update a graph layout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nodes:
                  type: array
                  items:
                    $ref: '#/components/schemas/GraphLayoutNode'
                edges:
                  type: array
                  items:
                    $ref: '#/components/schemas/GraphLayoutEdge'
                viewport:
                  $ref: '#/components/schemas/WorkflowViewport'
      responses:
        '200':
          description: Upserted layout
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GraphLayout'
    delete:
      operationId: deleteGraphLayout
      tags: [catalog-workflow-graph-layouts]
      summary: Delete a graph layout
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================================================
  # CATALOG-WORKFLOW — JSONata Assist
  # ==========================================================================
...
  /catalog-workflow/jsonata-assist/stream:
    post:
      operationId: streamJsonataExpression
      tags: [catalog-workflow-jsonata-assist]
      summary: Stream a JSONata expression generation via SSE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonataAssistRequest'
      responses:
        '200':
          description: SSE stream of generated expression
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  securitySchemes:
    ServiceToken:
      type: http
      scheme: bearer
      description: >
        A Roadie service token (`rst_…`) or a user JWT. Both are exchanged by the
        same middleware, so the wire contract is identical.


        Per-route scope requirements are deliberately not duplicated here. They
        live in `packages/e2e/tests/scopes/*.yaml`, which the scope-enforcement
        suite asserts covers every entry in `SCOPES` — that set is the
        machine-readable source, and a copy in this document would drift from it.


        One trap worth stating: a token carrying **no** `scopes` claim is granted
        every scope. A present array — including an empty one — restricts to
        exactly those. Scope a machine token explicitly, or it is a root
        credential.
  parameters:
...
    ExecutionId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          oneOf:
            - type: object
              properties:
                message:
                  type: string
                name:
                  type: string
            - type: string

    # --- Workflow Types ---
    WorkflowType:
      type: string
      enum: [data-ingestion]
...
    WorkflowViewport:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
        zoom:
          type: number
...
    NodeCategory:
      type: string
      enum: [trigger, source, transform, catalog, sink, control]
    HandleDefinition:
      type: object
      required: [id, label, type]
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [default, array, entity, location, any]
        required:
          type: boolean
    NodeTypeDefinition:
      type: object
      required: [type, category, label, description, icon, configSchema]
      properties:
        type:
          type: string
        category:
          $ref: '#/components/schemas/NodeCategory'
        label:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        configSchema: {}
        inputSchema: {}
        outputSchema: {}
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/HandleDefinition'
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/HandleDefinition'
        supportsDryRun:
          type: boolean
        tags:
          type: array
          items:
            type: string
        workflowTypes:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowType'

    # --- Graph Layout ---
    GraphLayoutNode:
      type: object
      properties:
        id:
          type: string
        datasourceId:
          type: string
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
    GraphLayoutEdge:
      type: object
      properties:
        id:
          type: string
        ruleId:
          type: string
        source:
          type: string
        target:
          type: string
        sourceHandle:
          type: string
        targetHandle:
          type: string
    GraphLayout:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/GraphLayoutNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphLayoutEdge'
        viewport:
          $ref: '#/components/schemas/WorkflowViewport'
        updatedBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # --- JSONata Assist ---
    JsonataAssistRequest:
      type: object
      required: [description]
      properties:
        description:
          type: string
        inputSample: {}
        transformType:
          type: string
          enum: [filter, map]

    # --- Integration Types ---
...